# Reduce memory usage by removing unnecessary regex