## What Changed
- Reduced memory usage(about 200-300MB 
in the data below) by removing regex compilation process when the search can be replaced with the fast search below.
  - https://github.com/Yamato-Security/hayabusa/pull/884
  - https://github.com/Yamato-Security/hayabusa/pull/883

- Changed pipe startswith/endswith matching from regex to https://github.com/Yamato-Security/hayabusa/pull/890 match.
  - This change didn't speed up matching, but it is changed for matching process refactoring

## Evidence

### Environment
- OS: Windows 10 Home edition
- Hard: Memory 16GB , Core 8, SSD, laptop

### Benchmark1
I ran a benchmark using [this procedure(6.1GB evtx)](https://github.com/Yamato-Security/hayabusa/issues/778#issuecomment-1296504766) and the results were as follows.

|Version| Elapsed time | Memory(peak)  | Events with hits / Total events | Output file size(bytes)|
|-------| ------------ | --------|--------|--------| 
| [before](https://github.com/Yamato-Security/hayabusa/tree/imporve-speed-by-using-starts-ends-with)  | 00:10:44.026| 4.7 GiB | 1,593,715 / 4,817,181 | 575085389 |
| [This PR](https://github.com/Yamato-Security/hayabusa/tree/improve-speed-by-reduce-regex-match-when-pipe)   | 00:10:42.195| 4.4 GiB | 1,593,715 / 4,817,181 | 575085385 | 

#### Console output

[before](https://github.com/Yamato-Security/hayabusa/tree/imporve-speed-by-using-starts-ends-with) 
```
PS C:\tmp\hayabusa-2.1.0-win-64-bit> .\hayabusa.exe csv-timeline -d ..\hayabusa-big-evtx\ -o 1.csv --debug
Results Summary:

Events with hits / Total events: 1,593,715 / 4,817,181 (Data reduction: 3,223,466 events (66.92%))

Total | Unique detections: 1,627,284 | 150
Total | Unique critical detections: 0 (0.00%) | 0 (0.00%)
Total | Unique high detections: 12,044 (0.74%) | 20 (13.33%)
Total | Unique medium detections: 11,118 (0.68%) | 38 (25.33%)
Total | Unique low detections: 1,053,623 (64.75%) | 42 (28.00%)
Total | Unique informational detections: 550,499 (33.83%) | 50 (33.33%)

...

Saved file: 1.csv (575.1 MB)
Elapsed time: 00:10:44.026
Rule Parse Processing Time: 00:00:20.191
Analysis Processing Time: 00:10:02.174
Output Processing Time: 00:00:21.658

Memory usage stats:
heap stats:    peak      total      freed    current       unit      count
  reserved:    4.7 GiB    4.7 GiB   83.0 MiB    4.6 GiB
 committed:    4.3 GiB   57.6 GiB   53.4 GiB    4.1 GiB
```

[This PR](https://github.com/Yamato-Security/hayabusa/tree/improve-speed-by-reduce-regex-match-when-pipe)
```
PS C:\tmp\hayabusa-2.1.0-win-64-bit> .\hayabusa.exe csv-timeline -d ..\hayabusa-big-evtx\ -o 1.csv --debug
Results Summary:

Events with hits / Total events: 1,593,715 / 4,817,181 (Data reduction: 3,223,466 events (66.92%))

Total | Unique detections: 1,627,284 | 150
Total | Unique critical detections: 0 (0.00%) | 0 (0.00%)
Total | Unique high detections: 12,044 (0.74%) | 20 (13.33%)
Total | Unique medium detections: 11,118 (0.68%) | 38 (25.33%)
Total | Unique low detections: 1,053,623 (64.75%) | 42 (28.00%)
Total | Unique informational detections: 550,499 (33.83%) | 50 (33.33%)
...

Saved file: 1.csv (575.1 MB)
Elapsed time: 00:10:42.195
Rule Parse Processing Time: 00:00:20.234
Analysis Processing Time: 00:10:00.446
Output Processing Time: 00:00:21.513

Memory usage stats:
heap stats:    peak      total      freed    current       unit      count
  reserved:    4.4 GiB    4.4 GiB   83.0 MiB    4.4 GiB
 committed:    4.0 GiB   65.0 GiB   61.2 GiB    3.8 GiB
```


### Benchmark2
I ran a benchmark using [hayabusa-sample-evtx](https://github.com/Yamato-Security/hayabusa-sample-evtx) and the results were as follows.

|Version| Elapsed time | Memory(peak)  | Events with hits / Total events | Output file size(bytes)|
|-------| ------------ | --------|--------|--------|
| [before](https://github.com/Yamato-Security/hayabusa/tree/imporve-speed-by-using-starts-ends-with)   | 00:00:16.158|  1.9 GiB | 19,606 / 47,458 | 16447790 | 
| [This PR](https://github.com/Yamato-Security/hayabusa/tree/improve-speed-by-reduce-regex-match-when-pipe)   | 00:00:15.422| 1.6 GiB| 19,606 / 47,458 | 16447790 | 

I would appreciate it if you could review🙏