@fukusuket 2023-01-25
@hitenkoku 
Thank you so much for quick review🙇 I fixed typo :) https://github.com/Yamato-Security/hayabusa/pull/894/commits/30d78d692a9626720b9a1c2918fae4a9ce5dacd6