@YamatoSecurity 2023-01-25
@fukusuket PRありがとうございます！
Macのベンチマークではメモリが300MB減っていますが、処理時間が少し増えています。
21:54.523 -> 22:26.437
１回目はメモリ使用 7.5GB -> 7.2GB、２回目は 7.5GB -> 7.4GBで、処理時間は二回目も同じ22:26でした。
明日Windowsの方でも試してみるので、少々お待ち下さい。