@fukusuket 2023-01-25
@YamatoSecurity 
ベンチマーク早速取得いただきありがとうございます🙇 
マッチ速度に影響しない想定の変更でしたが、すみません！どこかでデグレさせてしまっていそうです、確認いたします！
(またデータにより遅くなるパターンがありそうでしたら、こちらのPR一旦取り下げます！)