@YamatoSecurity 2023-01-25
@fukusuket Hummm, yea, it might just be my environment so I will check tonight on a different computer.